requirements.txt
loguru==0.7.3
python-dotenv==1.1.1
requests==2.32.4
vk-api==11.9.9
yt-dlp==2025.6.30
